fix(registrationPlatform): make sure this is updated for social users
